Currency 기능 구현(US고정 값)

+++ src/main/java/com/example/assignment/DTO/BookRequestDto.java
@@ -10,6 +10,7 @@ public class BookRequestDto {
     private int bookpage;
     private int age;
     private double price;
-    private int currency;
+    private double currency;
     private String authors;
+//    private List<Author> authorList;
 }
